Fixed weird test breakage
git-svn-id: http://agilefant.googlecode.com/svn/branches/Monza_restructure@2278 29784df9-99d5-7e28-3e6c-ac1830307ffc

+++ test/fi/hut/soberit/agilefant/util/ExactEstimateUtilsTest.java
@@ -9,43 +9,43 @@ public class ExactEstimateUtilsTest {
     @Test
     public void testValidHours() {
         assertEquals(60, ExactEstimateUtils.convertFromString("1h")
-                .getMinorUnits());
+                .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursWithWhitespace() {
         assertEquals(60, ExactEstimateUtils.convertFromString("  1h   ")
-                .getMinorUnits());
+                .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursWithDecimalsAsComma() {
         assertEquals(90, ExactEstimateUtils.convertFromString("1,5h")
-                .getMinorUnits());
+                .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursWithDecimalsAsPeriod() {
         assertEquals(90, ExactEstimateUtils.convertFromString("1.5h")
-                .getMinorUnits());
+                .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidMinutes() {
         assertEquals(15, ExactEstimateUtils.convertFromString("15min")
-                .getMinorUnits());
+                .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursAndMinutes() {
         assertEquals(105, ExactEstimateUtils.convertFromString("1.5h15min")
-                .getMinorUnits());
+                .getMinorUnits().longValue());
     }
 
     @Test
     public void testValidHoursAndMinutesWithWhitespace() {
         assertEquals(105, ExactEstimateUtils.convertFromString(
-                "   1.5h   15min        ").getMinorUnits());
+                "   1.5h   15min        ").getMinorUnits().longValue());
     }
 
     @Test(expected = IllegalArgumentException.class)
